fix(review): apply review suggestions (see details)
This includes:

- remove builders for mandatory arguments
- refactor the API to look more similar to the rest of the gizmo APIs
- split primitives module into 2d and 3d

Thanks for the thorough review Jondolf!

Authored-by: RobWalt <robwalter96@gmail.com>
