+++ crates/bevy_gizmos/src/primitives/mod.rs
@@ -1,0 +1,5 @@
+//! A module for rendering each of the 2D and 3D [`bevy_math::primitives`] with [`Gizmos`].
+
+pub mod dim2;
+pub mod dim3;
+pub(crate) mod helpers;
